thesis/livekit-worker/backend/main_api_service/requirements.txt
python-dotenv
livekit-agents
livekit
httpx
requests
cached_path
pydantic
scipy
livekit-plugins-silero
weaviate-client
pandas
duckdb
unidecode